chore: Rename package - user -> auth
