+++ src/main/java/seoultech/capstone/menjil/domain/auth/domain/SocialLoginType.java
@@ -1,0 +1,5 @@
+package seoultech.capstone.menjil.domain.auth.domain;
+
+public enum SocialLoginType {
+    GOOGLE, KAKAO
+}
