+++ src/main/java/seoultech/capstone/menjil/domain/auth/dto/KaKaoOAuthTokenDto.java
@@ -1,4 +1,4 @@
-package seoultech.capstone.menjil.domain.user.dto;
+package seoultech.capstone.menjil.domain.auth.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
