--- src/main/java/seoultech/capstone/menjil/domain/user/domain/SocialLoginType.java
@@ -1,5 +1,0 @@
-package seoultech.capstone.menjil.domain.user.domain;
-
-public enum SocialLoginType {
-    GOOGLE, KAKAO
-}
